+++ src/main/java/com/amigoscode/dto/AdminStatsResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -13,7 +14,7 @@ import java.util.Map;
 public class AdminStatsResponse {
 
     private long totalOrders;
-    private double totalRevenue;
+    private BigDecimal totalRevenue;
     private long totalBooks;
     private long totalUsers;
