+++ src/main/java/com/amigoscode/dto/CartItemResponse.java
@@ -2,14 +2,16 @@ package com.amigoscode.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemResponse {
-    private Long id; // ID của CartItem
+    private Long id;
     private Long bookId;
     private String bookTitle;
     private String bookAuthor;
-    private String imageUrl; // Ảnh bìa sách
-    private Double price;
+    private String imageUrl;
+    private BigDecimal price;
     private Integer quantity;
-    private Double subtotal; // = price * quantity
+    private BigDecimal subtotal; // = price × quantity
 }
