+++ src/main/java/com/amigoscode/repository/BookRepository.java
@@ -1,6 +1,7 @@
 package com.amigoscode.repository;
 
 import com.amigoscode.Entity.Book;
+import jakarta.persistence.LockModeType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,10 +9,10 @@ import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import jakarta.persistence.LockModeType;
-import java.util.Optional;
 
+import java.math.BigDecimal;
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface BookRepository extends JpaRepository<Book, Long> {
@@ -35,8 +36,8 @@ public interface BookRepository extends JpaRepository<Book, Long> {
             @Param("author") String author,
             @Param("isbn") String isbn,
             @Param("categoryName") String categoryName,
-            @Param("minPrice") Double minPrice,
-            @Param("maxPrice") Double maxPrice,
+            @Param("minPrice") BigDecimal minPrice,
+            @Param("maxPrice") BigDecimal maxPrice,
             Pageable pageable
     );
